requirements.txt
fastapi[standard]
uvicorn[standard]
google-adk
google-genai
python-dotenv
Langfuse